refactor: rename stateParam to state-param and update imports
